requirements.txt
Flask==1.1.2
Jinja2==2.11.2
Werkzeug==1.0.1
numpy==1.18.5
opencv-python==4.4.0.46
Pillow==8.0.1
sympy==1.6.2
tensorflow==2.3.1
protobuf==3.20.3
MarkupSafe==2.0.1

itsdangerous==1.1.0
click==7.1.2


